# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: OAuth2 Authentication Service
  description: |
    Enterprise-grade OAuth2 authentication service supporting Authorization Code Flow with PKCE
    and Client Credentials Flow for service-to-service authentication.

    This service provides secure token-based authentication with JWT access tokens,
    opaque refresh tokens, and comprehensive security features including rate limiting,
    CORS support, and Redis-based session management.
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1/auth
    description: Development server
  - url: https://auth-service.local/api/v1/auth
    description: Production server
security:
  - BearerAuth: []
  - ClientCredentials: []
paths:
...
  /oauth2/authorize:
    get:
      summary: OAuth2 authorization endpoint
      description: |
        Initiates the OAuth2 Authorization Code Flow with PKCE.
        This endpoint redirects the user to authenticate and authorize the client application.
      operationId: authorize
      tags:
        - OAuth2
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          description: Must be "code" for authorization code flow
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          description: The client identifier
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          description: The client redirect URI
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          description: The requested scope(s)
          schema:
            type: string
        - name: state
          in: query
          description: An opaque value to prevent CSRF attacks
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          description: PKCE code challenge
          schema:
            type: string
        - name: code_challenge_method
          in: query
          description: PKCE code challenge method
          schema:
            type: string
            enum: [S256]
            default: S256
      responses:
        '302':
          description: Redirect to client with authorization code or error
          headers:
            Location:
              description: Redirect URI with code or error parameters
              schema:
                type: string
                format: uri
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
    post:
      summary: OAuth2 authorization endpoint (POST)
      description: |
        Initiates the OAuth2 Authorization Code Flow with PKCE via POST.
        This endpoint redirects the user to authenticate and authorize the client application.
      operationId: authorizePost
      tags:
        - OAuth2
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                response_type:
                  type: string
                  enum: [code]
                client_id:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
                scope:
                  type: string
                state:
                  type: string
                code_challenge:
                  type: string
                code_challenge_method:
                  type: string
                  enum: [S256]
              required:
                - response_type
                - client_id
                - redirect_uri
                - code_challenge
      responses:
        '302':
          description: Redirect to client with authorization code or error
          headers:
            Location:
              description: Redirect URI with code or error parameters
              schema:
                type: string
                format: uri
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
  /oauth2/token:
    post:
      summary: OAuth2 token endpoint
      description: |
        Exchanges authorization codes for tokens or performs client credentials flow.
        Supports both Authorization Code Flow and Client Credentials Flow.
      operationId: token
      tags:
        - OAuth2
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              oneOf:
                - $ref: '#/components/schemas/AuthorizationCodeRequest'
                - $ref: '#/components/schemas/ClientCredentialsRequest'
                - $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
  /oauth2/introspect:
    post:
      summary: Token introspection endpoint
      description: Returns metadata about a token including its validity and associated claims
      operationId: introspect
      tags:
        - OAuth2
      security:
        - ClientCredentials: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IntrospectRequest'
      responses:
        '200':
          description: Token introspection response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
  /oauth2/revoke:
    post:
      summary: Token revocation endpoint
      description: Revokes an access token or refresh token
      operationId: revoke
      tags:
        - OAuth2
      security:
        - ClientCredentials: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RevokeRequest'
      responses:
        '200':
          description: Token revoked successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
...
  /oauth/clients:
    post:
      summary: Register OAuth2 client
      description: Registers a new OAuth2 client application (development/admin endpoint)
      operationId: registerClient
      tags:
        - OAuth2
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRegistrationRequest'
      responses:
        '201':
          description: Client registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientRegistrationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /oauth/clients/{client_id}:
    get:
      summary: Get OAuth2 client details
      description: Retrieves details of a registered OAuth2 client
      operationId: getClient
      tags:
        - OAuth2
      security: []
      parameters:
        - name: client_id
          in: path
          required: true
          description: The client identifier
          schema:
            type: string
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDetails'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /oauth/clients/{client_id}/secret:
    put:
      summary: Rotate OAuth2 client secret
      description: |
        Securely rotates an OAuth2 client's secret credential. The server generates a new
        cryptographically secure secret automatically. Requires the current secret for authentication.

        **Security Notes:**
        - New secret is generated server-side using cryptographically secure random (UUID)
        - The new secret is returned only once and cannot be retrieved again
        - Generated secrets are 36 characters (UUID format with dashes)
        - Client secrets are hashed using bcrypt (cost factor 12) before storage
        - All existing tokens remain valid after secret rotation

        **Use Cases:**
        - Regular security credential rotation (recommended every 90 days)
        - Suspected credential compromise
        - Compliance requirements for credential management

        **Best Practice:**
        This endpoint follows OAuth2 security best practices by generating secrets server-side,
        ensuring all credentials have strong, cryptographically secure entropy.
      operationId: updateClientSecret
      tags:
        - OAuth2
      security: []
      parameters:
        - name: client_id
          in: path
          required: true
          description: The client identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientSecretRotationRequest'
            example:
              current_secret: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Client secret rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSecretRotationResponse'
              example:
                client_id: "550e8400-e29b-41d4-a716-446655440000"
                new_secret: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                message: "Client secret rotated successfully. Store this secret securely - it will not be shown again."
        '400':
          description: Invalid request (e.g., missing current secret)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "validation_error"
                message: "current_secret is required"
                timestamp: "2025-11-10T00:56:45Z"
        '401':
          description: Unauthorized (invalid current secret or client not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
                message: "Failed to rotate secret: invalid credentials or client not found"  # pragma: allowlist secret
                timestamp: "2025-11-10T00:56:45Z"
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /oauth2/userinfo:
    get:
      summary: User information endpoint
      description: Returns information about the authenticated user
      operationId: userInfo
      tags:
        - OAuth2
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
    post:
      summary: User information endpoint (POST)
      description: Returns information about the authenticated user via POST
      operationId: userInfoPost
      tags:
        - OAuth2
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                access_token:
                  type: string
                  description: Access token (alternative to Authorization header)
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuth2Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ClientCredentials:
      type: http
      scheme: basic
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - error
        - message
        - timestamp
    OAuth2Error:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - access_denied
            - unsupported_response_type
            - server_error
            - temporarily_unavailable
        error_description:
          type: string
        error_uri:
          type: string
          format: uri
        state:
          type: string
      required:
        - error
    AuthorizationCodeRequest:
      type: object
      properties:
        grant_type:
          type: string
          enum: [authorization_code]
        code:
          type: string
        redirect_uri:
          type: string
          format: uri
        client_id:
          type: string
        code_verifier:
          type: string
      required:
        - grant_type
        - code
        - redirect_uri
        - client_id
        - code_verifier
    ClientCredentialsRequest:
      type: object
      properties:
        grant_type:
          type: string
          enum: [client_credentials]
        scope:
          type: string
      required:
        - grant_type
    RefreshTokenRequest:
      type: object
      properties:
        grant_type:
          type: string
          enum: [refresh_token]
        refresh_token:
          type: string
        scope:
          type: string
      required:
        - grant_type
        - refresh_token
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: The access token (JWT format)
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token lifetime in seconds
        refresh_token:
          type: string
          description: The refresh token (opaque format)
        scope:
          type: string
          description: The granted scope(s)
      required:
        - access_token
        - token_type
        - expires_in
    IntrospectRequest:
      type: object
      properties:
        token:
          type: string
        token_type_hint:
          type: string
          enum: [access_token, refresh_token]
      required:
        - token
    IntrospectResponse:
      type: object
      properties:
        active:
          type: boolean
        scope:
          type: string
        client_id:
          type: string
        sub:
          type: string
        exp:
          type: integer
        iat:
          type: integer
        aud:
          type: array
          items:
            type: string
      required:
        - active
    RevokeRequest:
      type: object
      properties:
        token:
          type: string
        token_type_hint:
          type: string
          enum: [access_token, refresh_token]
      required:
        - token
...
    UserInfo:
      type: object
      properties:
        sub:
          type: string
          description: Subject identifier
        email:
          type: string
          format: email
        name:
          type: string
        preferred_username:
          type: string
      required:
        - sub
...
    ClientRegistrationRequest:
      type: object
      properties:
        name:
          type: string
          description: Client application name
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
          description: Allowed redirect URIs
        scopes:
          type: array
          items:
            type: string
          description: Requested scopes
        grant_types:
          type: array
          items:
            type: string
          description: Supported grant types
      required:
        - name
        - redirect_uris
    ClientRegistrationResponse:
      type: object
      properties:
        client_id:
          type: string
        client_secret:
          type: string
        name:
          type: string
        redirect_uris:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        grant_types:
          type: array
          items:
            type: string
      required:
        - client_id
        - client_secret
        - name
        - redirect_uris
...
    ClientSecretRotationRequest:
      type: object
      properties:
        current_secret:
          type: string
          description: The current client secret for authentication (required to verify ownership)
          minLength: 1
          example: "550e8400-e29b-41d4-a716-446655440000"
      required:
        - current_secret
...